+++ src/main/java/com/asechallenge/inventorymanagement/dto/ProductResponseDTO.java
@@ -5,12 +5,12 @@ public class ProductResponseDTO {
     private String name;
     private String variant;
     private String description;
-    private Integer stockQuantity;
-    private Integer lowStockThreshold;
+    private Long stockQuantity;
+    private Long lowStockThreshold;
     private String sku;
 
     public ProductResponseDTO(String name, String variant, String description,
-                              Integer stockQuantity, Integer lowStockThreshold, String sku) {
+                              Long stockQuantity, Long lowStockThreshold, String sku) {
         this.name = name;
         this.variant = variant;
         this.description = description;
@@ -22,7 +22,7 @@ public class ProductResponseDTO {
     public String getName() { return name; }
     public String getVariant() { return variant; }
     public String getDescription() { return description; }
-    public Integer getStockQuantity() { return stockQuantity; }
-    public Integer getLowStockThreshold() { return lowStockThreshold; }
+    public Long getStockQuantity() { return stockQuantity; }
+    public Long getLowStockThreshold() { return lowStockThreshold; }
     public String getSku() { return sku; }
 }
